+++ src/imio/helpers/tests/test_content.py
@@ -24,6 +24,7 @@ from imio.helpers.content import richtextval
 from imio.helpers.content import safe_delattr
 from imio.helpers.content import safe_encode
 from imio.helpers.content import set_to_annotation
+from imio.helpers.content import sort_on_vocab_order
 from imio.helpers.content import uuidsToCatalogBrains
 from imio.helpers.content import uuidsToObjects
 from imio.helpers.content import uuidToCatalogBrain
@@ -455,7 +456,7 @@ class TestContentModule(IntegrationTestCase):
         self.assertEqual(object_ids(self.portal, ['Folder', 'ATFolder']),
                          ['folder', 'folder2'])
 
-    def test_user_fullname(self):
+    def test_get_user_fullname(self):
         self.assertIsNone(get_user_fullname(None))
         self.assertEqual(get_user_fullname("unknown_user", none_if_unfound=True), None)
         self.assertEqual(get_user_fullname("unknown_user"), "unknown_user")
@@ -479,5 +480,25 @@ class TestContentModule(IntegrationTestCase):
         self.assertTrue(getattr(self.portal.folder, 'dedana'))
         self.assertTrue(getattr(obj, 'dedana'))
         # base_getattr does not use acquisition
-        self.assertTrue(base_getattr(self.portal.folder, 'dedana'))
-        self.assertIsNone(base_getattr(obj, 'dedana'))
+        self.assertTrue(base_getattr(self.portal.folder, 'getImmediatelyAddableTypes'))
+        self.assertIsNone(base_getattr(obj, 'getImmediatelyAddableTypes'))
+
+    def test_sort_on_vocab_order(self):
+        self.assertEqual(
+            sort_on_vocab_order(('testingtype', 'Link', 'Document', 'File'),
+                                vocab_name='plone.app.vocabularies.PortalTypes'),
+            ['File', 'Link', 'Document', 'testingtype'])
+        vocab = get_vocab(self.portal, 'plone.app.vocabularies.PortalTypes')
+        self.assertEqual(
+            sort_on_vocab_order(('testingtype', 'Link', 'Document', 'File'),
+                                vocab=vocab),
+            ['File', 'Link', 'Document', 'testingtype'])
+        # not found values are moved to the end
+        self.assertEqual(
+            sort_on_vocab_order(('testingtype', 'unknown1', 'Link', 'Document', 'File'),
+                                vocab=vocab),
+            ['File', 'Link', 'Document', 'testingtype', 'unknown1'])
+        self.assertEqual(
+            sort_on_vocab_order(('testingtype', 'unknown2', 'Link', 'Document', 'File', 'unknown1'),
+                                vocab=vocab),
+            ['File', 'Link', 'Document', 'testingtype', 'unknown2', 'unknown1'])