+++ src/imio/helpers/tests/test_transmogrifier.py
@@ -1,8 +1,9 @@
 # -*- coding: utf-8 -*-
 from imio.helpers.testing import IntegrationTestCase
 from imio.helpers.transmogrifier import clean_value
-from imio.helpers.transmogrifier import correct_path
 from imio.helpers.transmogrifier import filter_keys
+from imio.helpers.transmogrifier import get_correct_id
+from imio.helpers.transmogrifier import get_correct_path
 from imio.helpers.transmogrifier import get_main_path
 from imio.helpers.transmogrifier import get_obj_from_path
 from imio.helpers.transmogrifier import key_val
@@ -32,20 +33,45 @@ class TestTesting(IntegrationTestCase):
         self.assertEqual(clean_value(u' strip  |  strip  ', isep=u'|'), u'strip|strip')
         self.assertEqual(clean_value(u'  \n strip  \n '), u'strip')
         self.assertEqual(clean_value(u' strip  \n "', strip=u' "'), u'strip')
-        self.assertEqual(clean_value(u' strip  \n "', patterns=[r'"']), u'strip')
-        self.assertEqual(clean_value(u' strip  \n "\'"', patterns=[r'^["\']+$']), u'strip')
+        self.assertEqual(clean_value(u' strip  \n "', patterns=[(r'"', u'')]), u'strip')
+        self.assertEqual(clean_value(u' strip  \n "\'"', patterns=[(r'^["\']+$', u'')]), u'strip')
         self.assertEqual(clean_value(u' strip  \n ', patterns=[]), u'strip')
         self.assertEqual(clean_value(u' strip  \n line2', patterns=[], osep=u'-'), u'strip-line2')
+        self.assertEqual(clean_value(u' stri    p  \n line  2', patterns=[(r'\s{2,}', u' ')], osep=u'-'),
+                         u'stri p-line 2')
 
-    def test_correct_path(self):
-        self.assertEquals(correct_path(self.portal, 'abcde'), 'abcde')
+    def test_get_correct_id(self):
+        self.assertEquals(get_correct_id(self.portal, 'abcde'), 'abcde')
         self.assertIn('folder', self.portal.objectIds())
-        self.assertEquals(correct_path(self.portal, 'folder'), 'folder-1')
-        self.assertEquals(correct_path(self.portal, 'folder/abcde'), 'folder/abcde')
+        self.assertEquals(get_correct_id(self.portal, 'folder'), 'folder-1')
+        self.assertEquals(get_correct_id(self.portal, 'folder', with_letter=True), 'folder-a')
+        self.assertEquals(get_correct_id(self.portal.folder, 'abcde'), 'abcde')
         self.portal.folder.invokeFactory('Document', id='abcde', title='Document')
-        self.assertEquals(correct_path(self.portal, 'folder/abcde'), 'folder/abcde-1')
+        self.assertEquals(get_correct_id(self.portal.folder, 'abcde'), 'abcde-1')
+        self.assertEquals(get_correct_id(self.portal.folder, 'abcde', True), 'abcde-a')
         self.portal.folder.invokeFactory('Document', id='abcde-1', title='Document')
-        self.assertEquals(correct_path(self.portal, 'folder/abcde'), 'folder/abcde-2')
+        self.assertEquals(get_correct_id(self.portal.folder, 'abcde'), 'abcde-2')
+
+    def test_get_correct_id_dic(self):
+        lst = ['a']
+        self.assertEquals(get_correct_id(lst, 'z'), 'z')
+        self.assertEquals(get_correct_id(lst, 'a'), 'a-1')
+        self.assertEquals(get_correct_id(lst, 'a', True), 'a-a')
+        lst.extend(['a-1', 'a-a'])
+        self.assertEquals(get_correct_id(lst, 'a'), 'a-2')
+        self.assertEquals(get_correct_id(lst, 'a', True), 'a-b')
+        lst.extend(['a-{}'.format(lt) for lt in 'bcdefghijklmnopqrstuvwxyz'])
+        self.assertEquals(get_correct_id(lst, 'a', True), 'a-aa')
+
+    def test_get_correct_path(self):
+        self.assertEquals(get_correct_path(self.portal, 'abcde'), 'abcde')
+        self.assertIn('folder', self.portal.objectIds())
+        self.assertEquals(get_correct_path(self.portal, 'folder'), 'folder-1')
+        self.assertEquals(get_correct_path(self.portal, 'folder/abcde'), 'folder/abcde')
+        self.portal.folder.invokeFactory('Document', id='abcde', title='Document')
+        self.assertEquals(get_correct_path(self.portal, 'folder/abcde'), 'folder/abcde-1')
+        self.portal.folder.invokeFactory('Document', id='abcde-1', title='Document')
+        self.assertEquals(get_correct_path(self.portal, 'folder/abcde'), 'folder/abcde-2')
 
     def test_filter_keys(self):
         dic = {'a': 1, 'b': 2, 'c': 3}
@@ -55,6 +81,10 @@ class TestTesting(IntegrationTestCase):
         self.assertListEqual([1, 2, 3], sorted(filter_keys(dic, ['a', 'b', 'c']).values()))
         self.assertListEqual(['a', 'c'], sorted(filter_keys(dic, ['a', 'c']).keys()))
         self.assertListEqual([1, 3], sorted(filter_keys(dic, ['a', 'c']).values()))
+        del dic['c']
+        self.assertListEqual(['a', 'b', 'c'], sorted(filter_keys(dic, ['a', 'b', 'c']).keys()))
+        self.assertListEqual([None, 1, 2], sorted(filter_keys(dic, ['a', 'b', 'c']).values()))
+        self.assertListEqual([1, 2, ''], sorted(filter_keys(dic, ['a', 'b', 'c'], unfound='').values()))
 
     def test_get_obj_from_path(self):
         folder = self.portal.folder
@@ -98,7 +128,8 @@ class TestTesting(IntegrationTestCase):
                          "value = '[1, 2, 3]'")
 
     def test_relative_path(self):
-        self.assertEquals(relative_path(self.portal, '/plone/directory'), 'directory')
+        self.assertEquals(relative_path(self.portal, '/plone/directory'), '/directory')
+        self.assertEquals(relative_path(self.portal, '/plone/directory', False), 'directory')
         self.assertEquals(relative_path(self.portal, '/alone/directory'), '/alone/directory')
 
     def test_split_text(self):
@@ -133,6 +164,13 @@ class TestTesting(IntegrationTestCase):
         # str_to_date(item, key, log_method, fmt='%Y/%m/%d', can_be_empty=True, as_date=True, **log_params):
         self.assertIsNone(str_to_date(dic, 'bad_key', logger))
         self.assertRaises(TypeError, str_to_date, dic, 'bad_key', logger, can_be_empty=False)
+        self.assertIsNone(str_to_date(dic, 1, logger, min_val=datetime.date(2023, 4, 1)))
+        self.assertEquals(dic['errors'], 1)
+        self.assertIsNone(str_to_date(dic, 1, logger, max_val=datetime.date(2022, 4, 1)))
+        self.assertEquals(dic['errors'], 2)
+        self.assertIsNone(str_to_date(dic, 1, logger, min_val=datetime.date(2023, 4, 1),
+                                      max_val=datetime.date(2022, 4, 1)))
+        self.assertEquals(dic['errors'], 3)
         ret = str_to_date(dic, 1, logger)
         self.assertIsInstance(ret, datetime.date)
         self.assertEqual(str(ret), '2023-03-03')
@@ -140,4 +178,4 @@ class TestTesting(IntegrationTestCase):
         self.assertIsInstance(ret, datetime.datetime)
         self.assertEqual(str(ret), '2023-03-03 09:29:00')
         ret = str_to_date(dic, 2, logger, fmt='%Y/%m/%d %M:%H', as_date=False)
-        self.assertEquals(dic['errors'], 1)
+        self.assertEquals(dic['errors'], 4)